devops_orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
requests==2.32.3
PyJWT==2.8.0
cryptography==42.0.8
